+++ tests/test_solver_general.py
@@ -100,6 +100,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         solve_params = {
             "n_iw": 10,
@@ -150,6 +151,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -265,6 +267,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -385,6 +388,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -503,6 +507,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -629,6 +634,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
                      h_loc_an_mat.reshape((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -749,6 +755,7 @@ class TestEDIpackSolverBathGeneral(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "general")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath + h_sc)
 
         # Part I: Initial solve()
         solve_params = {