+++ tests/test_solver_hybrid.py
@@ -90,6 +90,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         solve_params = {
             "n_iw": 10,
@@ -137,6 +138,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -241,6 +243,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -348,6 +351,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -452,6 +456,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -559,6 +564,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
                      h_loc_an_mat.reshape((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath)
 
         # Part I: Initial solve()
         solve_params = {
@@ -664,6 +670,7 @@ class TestEDIpackSolverBathHybrid(TestSolver):
         assert_equal(solver.hloc_an_mat, np.zeros((1, 1, self.norb, self.norb)))
         self.assertEqual(solver.bath.name, "hybrid")
         self.assertEqual(solver.bath.nbath, self.nbath)
+        self.assertEqual(solver.bath_op, h_bath + h_sc)
 
         # Part I: Initial solve()
         solve_params = {